Refactor test and use feature flag for e2e-tests
