+++ .github/workflows/rust.yml
@@ -47,13 +47,24 @@ jobs:
     - run: rustup update
     - run: rustup target add ${{ matrix.target }}
     - run: cargo clippy --all-targets --target ${{ matrix.target }}
-    # unit tests
-    - run: cargo test --bins --target ${{ matrix.target }}
-    # integration tests
-    # TODO: we should probably manage the container in the test itself
-    # TODO: do a proper functional check on the container startup instead of waiting 15 seconds
-    - run: docker run -d -p 8000:8000 stellar/quickstart:soroban-dev --standalone --enable-soroban-rpc --enable-core-artificially-accelerate-time-for-testing --protocol-version 20 && sleep 15
-    - run: cargo test --test 'integration_*' --target ${{ matrix.target }}
+    - run: cargo test --target ${{ matrix.target }}
+
+  e2e-tests:
+    strategy:
+      matrix:
+        include:
+          - os: ubuntu-latest
+            target: x86_64-unknown-linux-gnu
+    runs-on: ${{ matrix.os }}
+    steps:
+      - uses: actions/checkout@v3
+      - uses: stellar/actions/rust-cache@main
+      - run: rustup update
+      - run: rustup target add ${{ matrix.target }}
+      # TODO: we should probably manage the container in the test itself
+      # TODO: do a proper functional check on the container startup instead of waiting 15 seconds
+      - run: docker run -d -p 8000:8000 stellar/quickstart:soroban-dev --standalone --enable-soroban-rpc --enable-core-artificially-accelerate-time-for-testing --protocol-version 20 && sleep 15
+      - run: cargo test --features e2e_tests --test 'e2e*' --target ${{ matrix.target }}
 
   publish-dry-run:
     if: startsWith(github.head_ref, 'release/')
